Vista Solicitar nueva serie unificada y rediseñada Se ha unificado en un solo fichero la vista de buscar externamente y solicitar una serie nueva tanto de iOS como de Android Se ha implementado el nuevo diseño Se ha aplicado los nuevos Modal Se ha mejorado las funciones existentes, el código en general y los estilos
